Update index.html
I am updating this index.html so that it could directly access my project through github deployed page by the help of iframe which may help my project run on grand survey expedition as you can check yourself that it isn't running now. So please give it a try and merge this request.